content-security-policy/script-src/scriptnonce-specified-source.sub.html

<!DOCTYPE html>
<head>
    <title>CSP script with source and nonce specified.</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <meta http-equiv="Content-Security-Policy" content="script-src {{host}}:{{ports[http][0]}} 'nonce-nonce'"></meta>

    <script nonce="nonce">
        window.addEventListener('securitypolicyviolation', function(e) {
            alert_assert("Fail");
        });
    </script>


    <script nonce="nonce">
        var t_alert = async_test('CSP script with source and nonce specified.');
        var expected_alerts = ["PASS", "PASS"];

        function alert_assert(msg) {
            t_alert.step(function() {
                if (msg.match(/^FAIL/i)) {
                    assert_unreached(msg);
                    t_alert.done();
                }
                for (var i = 0; i < expected_alerts.length; i++) {
                    if (expected_alerts[i] == msg) {
                        assert_equals(expected_alerts[i], msg);
                        expected_alerts.splice(i, 1);
                        if (expected_alerts.length == 0) {
                            t_alert.done();
                        }
                        return;
                    }
                }
                assert_unreached('unexpected alert: ' + msg);
                t_alert.done();
            });
        }

    </script>

</head>
<body>
    <script nonce="noncynonce" src="http://{{host}}:{{ports[http][0]}}/content-security-policy/support/alert-pass.js"></script>
    <script nonce="nonce" src="http://{{host}}:{{ports[http][0]}}/content-security-policy/support/alert-pass.js"></script>
</body>
